extracted rom structure manipulation to a module
