+++ include/memory.h
@@ -3,54 +3,15 @@
 #include <stdbool.h>
 #include <inttypes.h>
 
+#include "rom.h"
 /*
  * biggest rom encountered so far is Pokemon Red-Blue 2-in-1.
  * increase this if no true.
  */
-#define ROM_BANK_SIZE 0x4000
 #define BIGGEST_ROM_SIZE 0x200000
 #define MEMORY_ADDRESS_SPACE 0x10000
 #define TOTAL_MEMORY (BIGGEST_ROM_SIZE + MEMORY_ADDRESS_SPACE - ROM_BANK_SIZE)
 #define EXTRA_ROM_BANKS_SIZE (TOTAL_MEMORY - MEMORY_ADDRESS_SPACE)
-
-#define MAX_TITLE_LENGTH 16
-#define LOGO_LENGTH 48
-#define HEADER_OFFSET_S 0x0100
-#define HEADER_OFFSET_E 0x014F
-
-#pragma pack(push, 1)
-
-struct rom_header {
-	uint32_t entry_point;
-	uint8_t logo[LOGO_LENGTH];
-	char title[MAX_TITLE_LENGTH];
-	uint16_t manufacturer_mode;
-	uint8_t cgb_flag;
-	uint8_t cartridge_type;
-	uint8_t rom_size;
-	uint8_t ram_size;
-	uint8_t dest_code; /* 00 JPN 01 N-JPN */
-	uint8_t old_licensee_code;
-	uint8_t game_version;
-	uint8_t header_checksum;
-	uint16_t glb_checksum;
-} __attribute__((__packed__));
-
-struct rom {
-	union {
-		uint8_t data[ROM_BANK_SIZE];
-		struct {
-			uint8_t padding[HEADER_OFFSET_S];
-			struct rom_header rom_header;
-		};
-	};
-} __attribute__((__packed__));
-
-#pragma pack(pop)
-
-int rom_init(struct rom *rom, struct rom *switchable_rom_bank,
-		uint8_t *extra_rom_banks, const char *filename);
-void rom_display_header(struct rom_header *rom_header);
 
 #pragma pack(push, 1)
